+++ instrument-repl/src/repl.rs
@@ -185,7 +185,12 @@ impl Repl {
         Ok(())
     }
 
-    fn handle_script_request(&mut self, file: &Path) -> Result<(bool, bool)> {
+    fn handle_script_request(
+        &mut self,
+        file: &Path,
+        save: bool,
+        run: bool,
+    ) -> Result<(bool, bool)> {
         let re = Regex::new(r"[^A-Za-z\d_]");
         let prompt = false;
         let command_written = true;
@@ -208,7 +213,7 @@ impl Repl {
                 let script_name = format!("kic_{result}");
 
                 self.inst
-                    .write_script(script_name.as_bytes(), contents.as_bytes(), false, true)?;
+                    .write_script(script_name.as_bytes(), contents.as_bytes(), save, run)?;
             }
             Err(err_msg) => {
                 unreachable!("Issue with regex creation: {}", err_msg.to_string());
@@ -433,7 +438,7 @@ impl Repl {
                                             .as_bytes(),
                                     )?;
                                     (prompt, command_written) =
-                                        self.handle_script_request(&file)?;
+                                        self.handle_script_request(&file, false, true)?;
                                 }
                                 SaveMethod::Buffers {
                                     names,
@@ -468,8 +473,9 @@ impl Repl {
                                 }
                             }
                         }
-                        Request::Script { file } => {
-                            (prompt, command_written) = self.handle_script_request(&file)?;
+                        Request::Script { file, save, run } => {
+                            (prompt, command_written) =
+                                self.handle_script_request(&file, save, run)?;
                         }
                         Request::TspLinkNodes { json_file } => {
                             self.set_lang_config_path(json_file.to_string_lossy().to_string());
@@ -834,6 +840,12 @@ impl Repl {
                         .required_unless_present("help")
                         .help("Path to the TSP script file to be sent to the instrument")
                 )
+                .arg(
+                    arg!(save: -s --save "Save the tsp script to non-volatile memory on the instrument").action(ArgAction::SetTrue)
+                )
+                .arg(
+                    arg!(run: -r --run <RUN_ENABLE> "Run the script after loading (defaults to true)").value_parser(value_parser!(bool)).default_value("true")
+                )
         )
         .subcommand(
             Command::new(".upgrade").about("Upgrade the firmware on the connected instrument")
@@ -957,7 +969,11 @@ impl Repl {
         let path = PathBuf::from(input.trim());
         if path.is_file() {
             trace!("Detected file path: {path:?}");
-            return Ok(Request::Script { file: path });
+            return Ok(Request::Script {
+                file: path,
+                save: false,
+                run: true,
+            });
         }
 
         if !Self::starts_with_command(input) {
@@ -1131,6 +1147,8 @@ impl Repl {
                             details: "expected file path, but none were provided".to_string(),
                         });
                     };
+                    let run: bool = *flags.get_one::<bool>("run").unwrap_or(&true);
+                    let save: bool = *flags.get_one::<bool>("save").unwrap_or(&false);
                     let file = PathBuf::from(file);
                     if !file.is_file() {
                         return Ok(Request::Usage(
@@ -1141,7 +1159,7 @@ impl Repl {
                             .to_string(),
                         ));
                     }
-                    Request::Script { file }
+                    Request::Script { file, save, run }
                 }
             },
             Some((".reset", flags)) => match flags.get_one::<bool>("help") {